requirements.txt
streamlit
langchain
ollama
python-dotenv
